Use explicit git materials in pipelines again
This reverts a recent change where we changed explicit git materials
with GoCD's option to use the current configrepo as material source.

This breaks for pipelines that have a non-master branch specified.
For these it would always use the master branch instead.

Also it turns out we don't need this change after all,
since we found an easier way to achieve what we want.

This is part of the 'restricting internet access on CI' epic that
BEEP is currently working on. Please talk to us if you have any questions.