Downgrade pinkie to supported version 2.0.4
Pinkie was (possibly maliciously) taken over and republished. Our cache cached
the "bad" version 2.0.5 even though it was removed from the official npm
registry, so it was installed in many places, including in this repo. This
commit downgrades to the official 2.0.4.

The v2.0.5 version was identical to 2.0.4 in code and configuration, so no
further action needs to be taken and nothing was exploited.