Update to use node.js LTS (currently 14) & npm 7
This updates to use node.js version 14, and npm 7. It uses the
nodejs-lts(-dev) runtime, which is currently the same as nodejs-14(-dev), but
will keep up to date with the Long Term Support version, reducing the need for
such changes in the future.

It also switches to use the new npm.localproxy.sc:42042 vpn-less registry.

This was a semi-automatically generated commit